requirements.txt
absl-py==0.8.0
aiocontextvars==0.2.2
albumentations==0.5.1
asn1crypto==0.24.0
astor==0.8.0
biwrap==0.1.6
certifi==2020.12.5
chardet==4.0.0
contextvars==2.4
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.0
h5py==2.10.0
idna==2.6
imageio==2.9.0
imgaug==0.4.0
immutables==0.14
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
loguru==0.5.3
Markdown==3.1.1
matplotlib==3.1.2
mock==3.0.5
networkx==2.5
numpy==1.16.4
opencv-python==4.1.2.30
opencv-python-headless==4.4.0.46
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.9.2
pycocotools==2.0.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.6
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
PyTurboJPEG==1.4.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
requests==2.25.1
scikit-image==0.17.2
scipy==1.5.2
SecretStorage==2.3.1
Shapely==1.7.1
six==1.11.0
ssh-import-id==5.7
tensorboard==1.13.1
tensorboardX==2.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-plot==0.3.2
termcolor==1.1.0
tifffile==2020.9.3
tqdm==4.46.0
urllib3==1.26.3
webcolors==1.11.1
Werkzeug==0.16.0
wrapt==1.11.2
